LayoutTests/imported/w3c/web-platform-tests/css/css-color/system-color-consistency-expected.txt


FAIL Property color value 'ButtonBorder' resolves to the same color as the border-color of a button assert_true: 'ButtonBorder' is a supported value for color. expected true got false
PASS Property color value 'ButtonFace' resolves to the same color as the background-color of a button
PASS Property color value 'ButtonText' resolves to the same color as text on a button
FAIL Property color value 'ButtonBorder' resolves to the same color as the border-color of a submit button assert_true: 'ButtonBorder' is a supported value for color. expected true got false
PASS Property color value 'ButtonFace' resolves to the same color as the background-color of a submit button
PASS Property color value 'ButtonText' resolves to the same color as text on a submit button
FAIL Property color value 'ButtonBorder' resolves to the same color as the border-color of a reset button assert_true: 'ButtonBorder' is a supported value for color. expected true got false
PASS Property color value 'ButtonFace' resolves to the same color as the background-color of a reset button
PASS Property color value 'ButtonText' resolves to the same color as text on a reset button
FAIL Property color value 'ButtonBorder' resolves to the same color as the border-color of a color picker assert_true: 'ButtonBorder' is a supported value for color. expected true got false
FAIL Property color value 'ButtonFace' resolves to the same color as the background-color of a color picker assert_equals: expected "rgb(255, 255, 255)" but got "rgb(192, 192, 192)"
FAIL Property color value 'ButtonText' resolves to the same color as text on a color picker assert_equals: expected "rgb(0, 0, 0)" but got "rgba(0, 0, 0, 0.847)"
PASS Property color value 'CanvasText' has the same color as the color of the html element
PASS Property color value 'Field' resolves to the same color as the background-color of a text field
FAIL Property color value 'FieldText' resolves to the same color as text on a text field assert_equals: expected "rgb(0, 0, 0)" but got "rgba(0, 0, 0, 0.847)"
PASS Property color value 'Field' resolves to the same color as the background-color of a password field
FAIL Property color value 'FieldText' resolves to the same color as text on a password field assert_equals: expected "rgb(0, 0, 0)" but got "rgba(0, 0, 0, 0.847)"
PASS Property color value 'Field' resolves to the same color as the background-color of a email field
FAIL Property color value 'FieldText' resolves to the same color as text on a email field assert_equals: expected "rgb(0, 0, 0)" but got "rgba(0, 0, 0, 0.847)"
PASS Property color value 'Field' resolves to the same color as the background-color of a number field
FAIL Property color value 'FieldText' resolves to the same color as text on a number field assert_equals: expected "rgb(0, 0, 0)" but got "rgba(0, 0, 0, 0.847)"
PASS Property color value 'Field' resolves to the same color as the background-color of a date field
FAIL Property color value 'FieldText' resolves to the same color as text on a date field assert_equals: expected "rgb(0, 0, 0)" but got "rgba(0, 0, 0, 0.847)"
PASS Property color value 'Field' resolves to the same color as the background-color of a text area
FAIL Property color value 'FieldText' resolves to the same color as text on a text area assert_equals: expected "rgb(0, 0, 0)" but got "rgba(0, 0, 0, 0.847)"
FAIL Property color value 'Mark' has the same color as the background-color of a mark element assert_true: 'Mark' is a supported value for color. expected true got false
FAIL Property color value 'MarkText' has the same color as the color of a mark element assert_true: 'MarkText' is a supported value for color. expected true got false
